Refactor request types
Request types conflated the action (request or invite) with the type of
the resource being added to a group (user/ws/catalog method), and also
enforced a specific type.

this results in a lot of duplicated code and poor extensibility.

Refactor the types into:
- a request type (invite/request)
- an arbitrary resource type (currently the application builder installs
workspace and catalogmethod, user is built in
- an arbitrary resource descriptor