Add a kafka based notifier
Tested manually for now.

To do:

Figure out how to bail cleanly when Kafka isn't reachable. Currently the
client continues trying to connect 1/s even when it goes out of scope.
If close() is called on the client with no timeout it hangs forever.

Tests

Javadoc

regex on allowable topic names

In the future, possibly support more client configuration keys