Revert group list view owner field to owner name
Was a data structure including the user fields and join date, but that
means in order to get that data & the member count the entire member
array has to be fetched from mongo. This precludes various optimizations
to return info that's not necessary in the list view, so reverting to
just the name to allow for mongo projection optimizations in the future.